Add `unstyled` option which skips inline styles on the wrapper div
And show in a new "eager" example module. This is how Wistia is likely to use our helpers (and likely other's who inject their own scripts for styling and sizing). Also port options to `snake_case`, as this is the HubL norm and how we desire our new tag's params to look